+++ bugsnag-plugin-react-native/src/test/java/com/bugsnag/android/ConfigSerializerTest.java
@@ -38,10 +38,9 @@ public class ConfigSerializerTest {
         assertTrue((Boolean) map.get("autoTrackSessions"));
         assertEquals("ALWAYS", map.get("sendThreads"));
 
-        String discardClassName = "com.example.DiscardClass";
-        Pattern pattern = Pattern.compile(discardClassName,Pattern.LITERAL);
-        Matcher matcher = pattern.matcher(map.get("discardClasses").toString());
-        assertTrue( matcher.find());
+        Matcher matcher = Pattern.compile("com.example.DiscardClass", Pattern.LITERAL)
+                .matcher(map.get("discardClasses").toString());
+        assertTrue(matcher.find());
 
         assertEquals(Collections.singleton("production"), map.get("enabledReleaseStages"));
         assertEquals(Collections.singleton("com.example"), map.get("projectPackages"));